+++ webapp/golang/app.go
@@ -1,99 +1,29 @@
 package main
 
 import (
-	crand "crypto/rand"
 	"fmt"
-	"html/template"
-	"io"
-	"io/ioutil"
 	"log"
 	"net/http"
-	"net/url"
 	"os"
-	"os/exec"
 	"path"
-	"path/filepath"
-	"regexp"
 	"strconv"
-	"strings"
-	"sync"
-	"time"
 
+	"github.com/catatsuy/private-isu/webapp/golang/handlers"
+	"github.com/catatsuy/private-isu/webapp/golang/utils"
 	"github.com/bradfitz/gomemcache/memcache"
 	gsm "github.com/bradleypeabody/gorilla-sessions-memcache"
 	"github.com/go-chi/chi/v5"
 	_ "github.com/go-sql-driver/mysql"
-	"github.com/gorilla/sessions"
 	"github.com/jmoiron/sqlx"
-	"golang.org/x/sync/errgroup"
 )
 
 var (
 	db    *sqlx.DB
 	store *gsm.MemcacheStore
-
-	// Static file cache
-	staticCache = make(map[string][]byte)
-	staticCacheMutex sync.RWMutex
 )
 
-const (
-	postsPerPage  = 20
-	ISO8601Format = "2006-01-02T15:04:05-07:00"
-	UploadLimit   = 10 * 1024 * 1024 // 10mb
-)
-
-type User struct {
-	ID          int       `db:"id"`
-	AccountName string    `db:"account_name"`
-	Passhash    string    `db:"passhash"`
-	Authority   int       `db:"authority"`
-	DelFlg      int       `db:"del_flg"`
-	CreatedAt   time.Time `db:"created_at"`
-}
-
-type Post struct {
-	ID           int       `db:"id"`
-	UserID       int       `db:"user_id"`
-	Imgdata      []byte    `db:"imgdata"`
-	Body         string    `db:"body"`
-	Mime         string    `db:"mime"`
-	CreatedAt    time.Time `db:"created_at"`
-	CommentCount int
-	Comments     []Comment
-	User         User
-	CSRFToken    string
-}
-
-type Comment struct {
-	ID        int       `db:"id"`
-	PostID    int       `db:"post_id"`
-	UserID    int       `db:"user_id"`
-	Comment   string    `db:"comment"`
-	CreatedAt time.Time `db:"created_at"`
-	User      User
-}
-
-// Async processing helpers
-type PostResult struct {
-	Post Post
-	Err  error
-}
-
-type CommentResult struct {
-	Comments []Comment
-	Count   int
-	Err     error
-}
-
-func init() {
-	memdAddr := os.Getenv("ISUCONP_MEMCACHED_ADDRESS")
-	if memdAddr == "" {
-		memdAddr = "localhost:11211"
-	}
-	memcacheClient := memcache.New(memdAddr)
-	store = gsm.NewMemcacheStore(memcacheClient, "iscogram_", []byte("sendagaya"))
-	log.SetFlags(log.Ldate | log.Ltime | log.Lshortfile)
+func getTemplPath(filename string) string {
+	return path.Join("templates", filename)
 }
 
 func dbInitialize() {
@@ -108,862 +38,6 @@ func dbInitialize() {
 	for _, sql := range sqls {
 		db.Exec(sql)
 	}
-}
-
-func tryLogin(accountName, password string) *User {
-	u := User{}
-	err := db.Get(&u, "SELECT * FROM users WHERE account_name = ? AND del_flg = 0", accountName)
-	if err != nil {
-		return nil
-	}
-
-	if calculatePasshash(u.AccountName, password) == u.Passhash {
-		return &u
-	} else {
-		return nil
-	}
-}
-
-func validateUser(accountName, password string) bool {
-	return regexp.MustCompile(`\A[0-9a-zA-Z_]{3,}\z`).MatchString(accountName) &&
-		regexp.MustCompile(`\A[0-9a-zA-Z_]{6,}\z`).MatchString(password)
-}
-
-// 今回のGo実装では言語側のエスケープの仕組みが使えないのでOSコマンドインジェクション対策できない
-// 取り急ぎPHPのescapeshellarg関数を参考に自前で実装
-// cf: http://jp2.php.net/manual/ja/function.escapeshellarg.php
-func escapeshellarg(arg string) string {
-	return "'" + strings.Replace(arg, "'", "'\\''", -1) + "'"
-}
-
-func digest(src string) string {
-	// opensslのバージョンによっては (stdin)= というのがつくので取る
-	out, err := exec.Command("/bin/bash", "-c", `printf "%s" `+escapeshellarg(src)+` | openssl dgst -sha512 | sed 's/^.*= //'`).Output()
-	if err != nil {
-		log.Print(err)
-		return ""
-	}
-
-	return strings.TrimSuffix(string(out), "\n")
-}
-
-func calculateSalt(accountName string) string {
-	return digest(accountName)
-}
-
-func calculatePasshash(accountName, password string) string {
-	return digest(password + ":" + calculateSalt(accountName))
-}
-
-func getSession(r *http.Request) *sessions.Session {
-	session, _ := store.Get(r, "isuconp-go.session")
-
-	return session
-}
-
-func getSessionUser(r *http.Request) User {
-	session := getSession(r)
-	uid, ok := session.Values["user_id"]
-	if !ok || uid == nil {
-		return User{}
-	}
-
-	u := User{}
-
-	err := db.Get(&u, "SELECT * FROM `users` WHERE `id` = ?", uid)
-	if err != nil {
-		return User{}
-	}
-
-	return u
-}
-
-func getFlash(w http.ResponseWriter, r *http.Request, key string) string {
-	session := getSession(r)
-	value, ok := session.Values[key]
-
-	if !ok || value == nil {
-		return ""
-	} else {
-		delete(session.Values, key)
-		session.Save(r, w)
-		return value.(string)
-	}
-}
-
-func makePosts(results []Post, csrfToken string, allComments bool) ([]Post, error) {
-	var posts []Post
-	postChan := make([]chan PostResult, len(results))
-	
-	// Start async processing for each post
-	for i := range results {
-		postChan[i] = getPostDataAsync(&results[i], csrfToken, allComments)
-	}
-	
-	// Collect results
-	for _, ch := range postChan {
-		result := <-ch
-		if result.Err != nil {
-			return nil, result.Err
-		}
-		
-		if result.Post.User.DelFlg == 0 {
-			posts = append(posts, result.Post)
-		}
-		if len(posts) >= postsPerPage {
-			break
-		}
-	}
-	
-	return posts, nil
-}
-
-func imageURL(p Post) string {
-	ext := ""
-	if p.Mime == "image/jpeg" {
-		ext = ".jpg"
-	} else if p.Mime == "image/png" {
-		ext = ".png"
-	} else if p.Mime == "image/gif" {
-		ext = ".gif"
-	}
-
-	return "/image/" + strconv.Itoa(p.ID) + ext
-}
-
-func isLogin(u User) bool {
-	return u.ID != 0
-}
-
-func getCSRFToken(r *http.Request) string {
-	session := getSession(r)
-	csrfToken, ok := session.Values["csrf_token"]
-	if !ok {
-		return ""
-	}
-	return csrfToken.(string)
-}
-
-func secureRandomStr(b int) string {
-	k := make([]byte, b)
-	if _, err := crand.Read(k); err != nil {
-		panic(err)
-	}
-	return fmt.Sprintf("%x", k)
-}
-
-func getTemplPath(filename string) string {
-	return path.Join("templates", filename)
-}
-
-func getInitialize(w http.ResponseWriter, r *http.Request) {
-	dbInitialize()
-	w.WriteHeader(http.StatusOK)
-}
-
-func getLogin(w http.ResponseWriter, r *http.Request) {
-	me := getSessionUser(r)
-
-	if isLogin(me) {
-		http.Redirect(w, r, "/", http.StatusFound)
-		return
-	}
-
-	template.Must(template.ParseFiles(
-		getTemplPath("layout.html"),
-		getTemplPath("login.html")),
-	).Execute(w, struct {
-		Me    User
-		Flash string
-	}{me, getFlash(w, r, "notice")})
-}
-
-func postLogin(w http.ResponseWriter, r *http.Request) {
-	if isLogin(getSessionUser(r)) {
-		http.Redirect(w, r, "/", http.StatusFound)
-		return
-	}
-
-	u := tryLogin(r.FormValue("account_name"), r.FormValue("password"))
-
-	if u != nil {
-		session := getSession(r)
-		session.Values["user_id"] = u.ID
-		session.Values["csrf_token"] = secureRandomStr(16)
-		session.Save(r, w)
-
-		http.Redirect(w, r, "/", http.StatusFound)
-	} else {
-		session := getSession(r)
-		session.Values["notice"] = "アカウント名かパスワードが間違っています"
-		session.Save(r, w)
-
-		http.Redirect(w, r, "/login", http.StatusFound)
-	}
-}
-
-func getRegister(w http.ResponseWriter, r *http.Request) {
-	if isLogin(getSessionUser(r)) {
-		http.Redirect(w, r, "/", http.StatusFound)
-		return
-	}
-
-	template.Must(template.ParseFiles(
-		getTemplPath("layout.html"),
-		getTemplPath("register.html")),
-	).Execute(w, struct {
-		Me    User
-		Flash string
-	}{User{}, getFlash(w, r, "notice")})
-}
-
-func postRegister(w http.ResponseWriter, r *http.Request) {
-	if isLogin(getSessionUser(r)) {
-		http.Redirect(w, r, "/", http.StatusFound)
-		return
-	}
-	
-	accountName, password := r.FormValue("account_name"), r.FormValue("password")
-	
-	validated := validateUser(accountName, password)
-	if !validated {
-		session := getSession(r)
-		session.Values["notice"] = "アカウント名は3文字以上、パスワードは6文字以上である必要があります"
-		session.Save(r, w)
-		
-		http.Redirect(w, r, "/register", http.StatusFound)
-		return
-	}
-	
-	exists := 0
-	// Start password hash calculation while checking user existence
-	passhashChan := calculatePasshashAsync(accountName, password)
-	
-	err := db.Get(&exists, "SELECT 1 FROM users WHERE `account_name` = ?", accountName)
-	if err == nil && exists == 1 {
-		session := getSession(r)
-		session.Values["notice"] = "アカウント名がすでに使われています"
-		session.Save(r, w)
-		
-		http.Redirect(w, r, "/register", http.StatusFound)
-		return
-	}
-	
-	// Wait for password hash
-	passhash := <-passhashChan
-	
-	query := "INSERT INTO `users` (`account_name`, `passhash`) VALUES (?,?)"
-	result, err := db.Exec(query, accountName, passhash)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-	
-	session := getSession(r)
-	uid, err := result.LastInsertId()
-	if err != nil {
-		log.Print(err)
-		return
-	}
-	session.Values["user_id"] = uid
-	session.Values["csrf_token"] = secureRandomStr(16)
-	session.Save(r, w)
-	
-	http.Redirect(w, r, "/", http.StatusFound)
-}
-
-func getLogout(w http.ResponseWriter, r *http.Request) {
-	session := getSession(r)
-	delete(session.Values, "user_id")
-	session.Options = &sessions.Options{MaxAge: -1}
-	session.Save(r, w)
-
-	http.Redirect(w, r, "/", http.StatusFound)
-}
-
-func getIndex(w http.ResponseWriter, r *http.Request) {
-	me := getSessionUser(r)
-
-	results := []Post{}
-
-	err := db.Select(&results, "SELECT p.id, p.user_id, p.body, p.mime, p.created_at FROM posts p INNER JOIN users u ON p.user_id = u.id WHERE u.del_flg = 0 ORDER BY p.created_at DESC LIMIT ?", postsPerPage)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	posts, err := makePosts(results, getCSRFToken(r), false)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	fmap := template.FuncMap{
-		"imageURL": imageURL,
-	}
-
-	template.Must(template.New("layout.html").Funcs(fmap).ParseFiles(
-		getTemplPath("layout.html"),
-		getTemplPath("index.html"),
-		getTemplPath("posts.html"),
-		getTemplPath("post.html"),
-	)).Execute(w, struct {
-		Posts     []Post
-		Me        User
-		CSRFToken string
-		Flash     string
-	}{posts, me, getCSRFToken(r), getFlash(w, r, "notice")})
-}
-
-func getAccountName(w http.ResponseWriter, r *http.Request) {
-	accountName := r.PathValue("accountName")
-	user := User{}
-
-	err := db.Get(&user, "SELECT * FROM `users` WHERE `account_name` = ? AND `del_flg` = 0", accountName)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	if user.ID == 0 {
-		w.WriteHeader(http.StatusNotFound)
-		return
-	}
-
-	results := []Post{}
-
-	err = db.Select(&results, "SELECT `id`, `user_id`, `body`, `mime`, `created_at` FROM `posts` WHERE `user_id` = ? ORDER BY `created_at` DESC", user.ID)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	posts, err := makePosts(results, getCSRFToken(r), false)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	commentCount := 0
-	err = db.Get(&commentCount, "SELECT COUNT(*) AS count FROM `comments` WHERE `user_id` = ?", user.ID)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	postIDs := []int{}
-	err = db.Select(&postIDs, "SELECT `id` FROM `posts` WHERE `user_id` = ?", user.ID)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-	postCount := len(postIDs)
-
-	commentedCount := 0
-	if postCount > 0 {
-		s := []string{}
-		for range postIDs {
-			s = append(s, "?")
-		}
-		placeholder := strings.Join(s, ", ")
-
-		// convert []int -> []interface{}
-		args := make([]interface{}, len(postIDs))
-		for i, v := range postIDs {
-			args[i] = v
-		}
-
-		err = db.Get(&commentedCount, "SELECT COUNT(*) AS count FROM `comments` WHERE `post_id` IN ("+placeholder+")", args...)
-		if err != nil {
-			log.Print(err)
-			return
-		}
-	}
-
-	me := getSessionUser(r)
-
-	fmap := template.FuncMap{
-		"imageURL": imageURL,
-	}
-
-	template.Must(template.New("layout.html").Funcs(fmap).ParseFiles(
-		getTemplPath("layout.html"),
-		getTemplPath("user.html"),
-		getTemplPath("posts.html"),
-		getTemplPath("post.html"),
-	)).Execute(w, struct {
-		Posts          []Post
-		User           User
-		PostCount      int
-		CommentCount   int
-		CommentedCount int
-		Me             User
-	}{posts, user, postCount, commentCount, commentedCount, me})
-}
-
-func getPosts(w http.ResponseWriter, r *http.Request) {
-	m, err := url.ParseQuery(r.URL.RawQuery)
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Print(err)
-		return
-	}
-	maxCreatedAt := m.Get("max_created_at")
-	if maxCreatedAt == "" {
-		return
-	}
-
-	t, err := time.Parse(ISO8601Format, maxCreatedAt)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	results := []Post{}
-	err = db.Select(&results, "SELECT p.id, p.user_id, p.body, p.mime, p.created_at FROM posts p INNER JOIN users u ON p.user_id = u.id WHERE u.del_flg = 0 AND p.created_at <= ? ORDER BY p.created_at DESC LIMIT ?", t.Format(ISO8601Format), postsPerPage)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	posts, err := makePosts(results, getCSRFToken(r), false)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	if len(posts) == 0 {
-		w.WriteHeader(http.StatusNotFound)
-		return
-	}
-
-	fmap := template.FuncMap{
-		"imageURL": imageURL,
-	}
-
-	template.Must(template.New("posts.html").Funcs(fmap).ParseFiles(
-		getTemplPath("posts.html"),
-		getTemplPath("post.html"),
-	)).Execute(w, posts)
-}
-
-func getPostsID(w http.ResponseWriter, r *http.Request) {
-	pidStr := r.PathValue("id")
-	pid, err := strconv.Atoi(pidStr)
-	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		return
-	}
-
-	results := []Post{}
-	err = db.Select(&results, "SELECT * FROM `posts` WHERE `id` = ?", pid)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	posts, err := makePosts(results, getCSRFToken(r), true)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	if len(posts) == 0 {
-		w.WriteHeader(http.StatusNotFound)
-		return
-	}
-
-	p := posts[0]
-
-	me := getSessionUser(r)
-
-	fmap := template.FuncMap{
-		"imageURL": imageURL,
-	}
-
-	template.Must(template.New("layout.html").Funcs(fmap).ParseFiles(
-		getTemplPath("layout.html"),
-		getTemplPath("post_id.html"),
-		getTemplPath("post.html"),
-	)).Execute(w, struct {
-		Post Post
-		Me   User
-	}{p, me})
-}
-
-func postIndex(w http.ResponseWriter, r *http.Request) {
-	me := getSessionUser(r)
-	if !isLogin(me) {
-		http.Redirect(w, r, "/login", http.StatusFound)
-		return
-	}
-
-	if r.FormValue("csrf_token") != getCSRFToken(r) {
-		w.WriteHeader(http.StatusUnprocessableEntity)
-		return
-	}
-
-	file, header, err := r.FormFile("file")
-	if err != nil {
-		session := getSession(r)
-		session.Values["notice"] = "画像が必須です"
-		session.Save(r, w)
-
-		http.Redirect(w, r, "/", http.StatusFound)
-		return
-	}
-
-	mime := ""
-	if file != nil {
-		// 投稿のContent-Typeからファイルのタイプを決定する
-		contentType := header.Header["Content-Type"][0]
-		if strings.Contains(contentType, "jpeg") {
-			mime = "image/jpeg"
-		} else if strings.Contains(contentType, "png") {
-			mime = "image/png"
-		} else if strings.Contains(contentType, "gif") {
-			mime = "image/gif"
-		} else {
-			session := getSession(r)
-			session.Values["notice"] = "投稿できる画像形式はjpgとpngとgifだけです"
-			session.Save(r, w)
-
-			http.Redirect(w, r, "/", http.StatusFound)
-			return
-		}
-	}
-
-	filedata, err := io.ReadAll(file)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	if len(filedata) > UploadLimit {
-		session := getSession(r)
-		session.Values["notice"] = "ファイルサイズが大きすぎます"
-		session.Save(r, w)
-
-		http.Redirect(w, r, "/", http.StatusFound)
-		return
-	}
-
-	query := "INSERT INTO `posts` (`user_id`, `mime`, `imgdata`, `body`) VALUES (?,?,?,?)"
-	result, err := db.Exec(
-		query,
-		me.ID,
-		mime,
-		filedata,
-		r.FormValue("body"),
-	)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	pid, err := result.LastInsertId()
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	http.Redirect(w, r, "/posts/"+strconv.FormatInt(pid, 10), http.StatusFound)
-}
-
-func getImage(w http.ResponseWriter, r *http.Request) {
-	pidStr := r.PathValue("id")
-	pid, err := strconv.Atoi(pidStr)
-	if err != nil {
-		w.WriteHeader(http.StatusNotFound)
-		return
-	}
-
-	post := Post{}
-	err = db.Get(&post, "SELECT * FROM `posts` WHERE `id` = ?", pid)
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	ext := r.PathValue("ext")
-
-	if ext == "jpg" && post.Mime == "image/jpeg" ||
-		ext == "png" && post.Mime == "image/png" ||
-		ext == "gif" && post.Mime == "image/gif" {
-		w.Header().Set("Content-Type", post.Mime)
-		_, err := w.Write(post.Imgdata)
-		if err != nil {
-			log.Print(err)
-			return
-		}
-		return
-	}
-
-	w.WriteHeader(http.StatusNotFound)
-}
-
-func postComment(w http.ResponseWriter, r *http.Request) {
-	me := getSessionUser(r)
-	if !isLogin(me) {
-		http.Redirect(w, r, "/login", http.StatusFound)
-		return
-	}
-
-	if r.FormValue("csrf_token") != getCSRFToken(r) {
-		w.WriteHeader(http.StatusUnprocessableEntity)
-		return
-	}
-
-	postID, err := strconv.Atoi(r.FormValue("post_id"))
-	if err != nil {
-		log.Print("post_idは整数のみです")
-		return
-	}
-
-	query := "INSERT INTO `comments` (`post_id`, `user_id`, `comment`) VALUES (?,?,?)"
-	_, err = db.Exec(query, postID, me.ID, r.FormValue("comment"))
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	http.Redirect(w, r, fmt.Sprintf("/posts/%d", postID), http.StatusFound)
-}
-
-func getAdminBanned(w http.ResponseWriter, r *http.Request) {
-	me := getSessionUser(r)
-	if !isLogin(me) {
-		http.Redirect(w, r, "/", http.StatusFound)
-		return
-	}
-
-	if me.Authority == 0 {
-		w.WriteHeader(http.StatusForbidden)
-		return
-	}
-
-	users := []User{}
-	err := db.Select(&users, "SELECT * FROM `users` WHERE `authority` = 0 AND `del_flg` = 0 ORDER BY `created_at` DESC")
-	if err != nil {
-		log.Print(err)
-		return
-	}
-
-	template.Must(template.ParseFiles(
-		getTemplPath("layout.html"),
-		getTemplPath("banned.html")),
-	).Execute(w, struct {
-		Users     []User
-		Me        User
-		CSRFToken string
-	}{users, me, getCSRFToken(r)})
-}
-
-func postAdminBanned(w http.ResponseWriter, r *http.Request) {
-	me := getSessionUser(r)
-	if !isLogin(me) {
-		http.Redirect(w, r, "/", http.StatusFound)
-		return
-	}
-	
-	if me.Authority == 0 {
-		w.WriteHeader(http.StatusForbidden)
-		return
-	}
-	
-	if r.FormValue("csrf_token") != getCSRFToken(r) {
-		w.WriteHeader(http.StatusUnprocessableEntity)
-		return
-	}
-	
-	err := r.ParseForm()
-	if err != nil {
-		log.Print(err)
-		return
-	}
-	
-	// Process bans in parallel
-	eg := errgroup.Group{}
-	for _, id := range r.Form["uid[]"] {
-		id := id // capture loop variable
-		eg.Go(func() error {
-			_, err := db.Exec("UPDATE `users` SET `del_flg` = ? WHERE `id` = ?", 1, id)
-			return err
-		})
-	}
-	
-	if err := eg.Wait(); err != nil {
-		log.Print(err)
-		return
-	}
-	
-	http.Redirect(w, r, "/admin/banned", http.StatusFound)
-}
-
-// Initialize static file cache
-func initStaticCache() error {
-	staticPaths := []string{
-		"../public/css",
-		"../public/js",
-		"../public/img",
-	}
-
-	for _, dir := range staticPaths {
-		err := filepath.Walk(dir, func(path string, info os.FileInfo, err error) error {
-			if err != nil {
-				return err
-			}
-			if !info.IsDir() {
-				data, err := ioutil.ReadFile(path)
-				if err != nil {
-					return err
-				}
-				relativePath := strings.TrimPrefix(path, "../public")
-				staticCacheMutex.Lock()
-				staticCache[relativePath] = data
-				staticCacheMutex.Unlock()
-			}
-			return nil
-		})
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func serveStaticFile(w http.ResponseWriter, r *http.Request) {
-	path := r.URL.Path
-	
-	// Try to serve from cache first
-	staticCacheMutex.RLock()
-	data, exists := staticCache[path]
-	staticCacheMutex.RUnlock()
-	
-	if exists {
-		// Set cache headers
-		w.Header().Set("Cache-Control", "public, max-age=2592000") // 30 days
-		w.Header().Set("Expires", time.Now().Add(30*24*time.Hour).Format(http.TimeFormat))
-		
-		// Determine content type
-		ext := filepath.Ext(path)
-		switch ext {
-		case ".css":
-			w.Header().Set("Content-Type", "text/css")
-		case ".js":
-			w.Header().Set("Content-Type", "application/javascript")
-		case ".jpg", ".jpeg":
-			w.Header().Set("Content-Type", "image/jpeg")
-		case ".png":
-			w.Header().Set("Content-Type", "image/png")
-		case ".gif":
-			w.Header().Set("Content-Type", "image/gif")
-		case ".ico":
-			w.Header().Set("Content-Type", "image/x-icon")
-		}
-		
-		w.Write(data)
-		return
-	}
-	
-	// Fallback to standard file server
-	http.FileServer(http.Dir("../public")).ServeHTTP(w, r)
-}
-
-// Async processing helpers
-func getPostDataAsync(p *Post, csrfToken string, allComments bool) chan PostResult {
-	resultChan := make(chan PostResult, 1)
-	
-	go func() {
-		defer close(resultChan)
-		
-		// Get comment count and comments in parallel
-		commentChan := getCommentsAsync(p.ID, allComments)
-		
-		// Get user info while waiting for comments
-		err := db.Get(&p.User, "SELECT * FROM `users` WHERE `id` = ?", p.UserID)
-		if err != nil {
-			resultChan <- PostResult{Err: err}
-			return
-		}
-		
-		// Wait for comment results
-		commentResult := <-commentChan
-		if commentResult.Err != nil {
-			resultChan <- PostResult{Err: commentResult.Err}
-			return
-		}
-		
-		p.Comments = commentResult.Comments
-		p.CommentCount = commentResult.Count
-		p.CSRFToken = csrfToken
-		
-		resultChan <- PostResult{Post: *p}
-	}()
-	
-	return resultChan
-}
-
-func getCommentsAsync(postID int, allComments bool) chan CommentResult {
-	resultChan := make(chan CommentResult, 1)
-	
-	go func() {
-		defer close(resultChan)
-		
-		var count int
-		err := db.Get(&count, "SELECT COUNT(*) AS `count` FROM `comments` WHERE `post_id` = ?", postID)
-		if err != nil {
-			resultChan <- CommentResult{Err: err}
-			return
-		}
-		
-		query := "SELECT * FROM `comments` WHERE `post_id` = ? ORDER BY `created_at` DESC"
-		if !allComments {
-			query += " LIMIT 3"
-		}
-		
-		var comments []Comment
-		err = db.Select(&comments, query, postID)
-		if err != nil {
-			resultChan <- CommentResult{Err: err}
-			return
-		}
-		
-		// Get user info for each comment in parallel
-		eg := errgroup.Group{}
-		for i := range comments {
-			i := i // capture loop variable
-			eg.Go(func() error {
-				return db.Get(&comments[i].User, "SELECT * FROM `users` WHERE `id` = ?", comments[i].UserID)
-			})
-		}
-		
-		if err := eg.Wait(); err != nil {
-			resultChan <- CommentResult{Err: err}
-			return
-		}
-		
-		// Reverse comments order
-		for i, j := 0, len(comments)-1; i < j; i, j = i+1, j-1 {
-			comments[i], comments[j] = comments[j], comments[i]
-		}
-		
-		resultChan <- CommentResult{Comments: comments, Count: count}
-	}()
-	
-	return resultChan
-}
-
-// Async password hashing
-func calculatePasshashAsync(accountName, password string) chan string {
-	resultChan := make(chan string, 1)
-	
-	go func() {
-		defer close(resultChan)
-		resultChan <- digest(password + ":" + calculateSalt(accountName))
-	}()
-	
-	return resultChan
 }
 
 func main() {
@@ -1004,33 +78,35 @@ func main() {
 	}
 	defer db.Close()
 
-	// Initialize static file cache
-	if err := initStaticCache(); err != nil {
-		log.Printf("Warning: Failed to initialize static cache: %v", err)
+	memdAddr := os.Getenv("ISUCONP_MEMCACHED_ADDRESS")
+	if memdAddr == "" {
+		memdAddr = "localhost:11211"
 	}
+	memcacheClient := memcache.New(memdAddr)
+	store = gsm.NewMemcacheStore(memcacheClient, "iscogram_", []byte("sendagaya"))
+
+	// Initialize handlers and utils
+	handlers.InitHandlers(db, store)
+	utils.InitSession(db, store)
 
 	r := chi.NewRouter()
 
-	// Static file routes
-	r.Get("/css/*", serveStaticFile)
-	r.Get("/js/*", serveStaticFile)
-	r.Get("/img/*", serveStaticFile)
-	r.Get("/favicon.ico", serveStaticFile)
+	// Routes
+	r.Get("/initialize", dbInitialize)
+	r.Get("/login", handlers.GetLogin)
+	r.Post("/login", handlers.PostLogin)
+	r.Get("/register", handlers.GetRegister)
+	r.Post("/register", handlers.PostRegister)
+	r.Get("/logout", handlers.GetLogout)
+	r.Get("/", handlers.GetIndex)
+	r.Post("/", handlers.PostIndex)
+	r.Get("/posts", handlers.GetPosts)
+	r.Get("/posts/{id}", handlers.GetPostsID)
+	r.Get("/image/{id}.{ext}", handlers.GetImage)
+	r.Post("/comment", handlers.PostComment)
+	r.Get("/admin/banned", handlers.GetAdminBanned)
+	r.Post("/admin/banned", handlers.PostAdminBanned)
+	r.Get(`/@{accountName:[a-zA-Z]+}`, handlers.GetAccountName)
 
-	r.Get("/initialize", getInitialize)
-	r.Get("/login", getLogin)
-	r.Post("/login", postLogin)
-	r.Get("/register", getRegister)
-	r.Post("/register", postRegister)
-	r.Get("/logout", getLogout)
-	r.Get("/", getIndex)
-	r.Get("/posts", getPosts)
-	r.Get("/posts/{id}", getPostsID)
-	r.Post("/", postIndex)
-	r.Get("/image/{id}.{ext}", getImage)
-	r.Post("/comment", postComment)
-	r.Get("/admin/banned", getAdminBanned)
-	r.Post("/admin/banned", postAdminBanned)
-	r.Get(`/@{accountName:[a-zA-Z]+}`, getAccountName)
 	log.Fatal(http.ListenAndServe(":8080", r))
 }
