+++ webapp/golang/handlers/auth.go
@@ -1,0 +1,139 @@
+package handlers
+
+import (
+	"log"
+	"net/http"
+	"regexp"
+
+	"github.com/gorilla/sessions"
+	"github.com/jmoiron/sqlx"
+)
+
+var (
+	db    *sqlx.DB
+	store sessions.Store
+)
+
+func InitHandlers(database *sqlx.DB, sessionStore sessions.Store) {
+	db = database
+	store = sessionStore
+}
+
+func validateUser(accountName, password string) bool {
+	return regexp.MustCompile(`\A[0-9a-zA-Z_]{3,}\z`).MatchString(accountName) &&
+		regexp.MustCompile(`\A[0-9a-zA-Z_]{6,}\z`).MatchString(password)
+}
+
+func GetLogin(w http.ResponseWriter, r *http.Request) {
+	me := GetSessionUser(r)
+
+	if IsLogin(me) {
+		http.Redirect(w, r, "/", http.StatusFound)
+		return
+	}
+
+	template.Must(template.ParseFiles(
+		getTemplPath("layout.html"),
+		getTemplPath("login.html")),
+	).Execute(w, struct {
+		Me    User
+		Flash string
+	}{me, GetFlash(w, r, "notice")})
+}
+
+func PostLogin(w http.ResponseWriter, r *http.Request) {
+	if IsLogin(GetSessionUser(r)) {
+		http.Redirect(w, r, "/", http.StatusFound)
+		return
+	}
+
+	u := tryLogin(r.FormValue("account_name"), r.FormValue("password"))
+
+	if u != nil {
+		session := GetSession(r)
+		session.Values["user_id"] = u.ID
+		session.Values["csrf_token"] = SecureRandomStr(16)
+		session.Save(r, w)
+
+		http.Redirect(w, r, "/", http.StatusFound)
+	} else {
+		session := GetSession(r)
+		session.Values["notice"] = "アカウント名かパスワードが間違っています"
+		session.Save(r, w)
+
+		http.Redirect(w, r, "/login", http.StatusFound)
+	}
+}
+
+func GetRegister(w http.ResponseWriter, r *http.Request) {
+	if IsLogin(GetSessionUser(r)) {
+		http.Redirect(w, r, "/", http.StatusFound)
+		return
+	}
+
+	template.Must(template.ParseFiles(
+		getTemplPath("layout.html"),
+		getTemplPath("register.html")),
+	).Execute(w, struct {
+		Me    User
+		Flash string
+	}{User{}, GetFlash(w, r, "notice")})
+}
+
+func PostRegister(w http.ResponseWriter, r *http.Request) {
+	if IsLogin(GetSessionUser(r)) {
+		http.Redirect(w, r, "/", http.StatusFound)
+		return
+	}
+
+	accountName, password := r.FormValue("account_name"), r.FormValue("password")
+
+	validated := validateUser(accountName, password)
+	if !validated {
+		session := GetSession(r)
+		session.Values["notice"] = "アカウント名は3文字以上、パスワードは6文字以上である必要があります"
+		session.Save(r, w)
+
+		http.Redirect(w, r, "/register", http.StatusFound)
+		return
+	}
+
+	exists := 0
+	err := db.Get(&exists, "SELECT 1 FROM users WHERE `account_name` = ?", accountName)
+	if err == nil && exists == 1 {
+		session := GetSession(r)
+		session.Values["notice"] = "アカウント名がすでに使われています"
+		session.Save(r, w)
+
+		http.Redirect(w, r, "/register", http.StatusFound)
+		return
+	}
+
+	query := "INSERT INTO `users` (`account_name`, `passhash`) VALUES (?,?)"
+	result, err := db.Exec(query, accountName, CalculatePasshash(accountName, password))
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	session := GetSession(r)
+	uid, err := result.LastInsertId()
+	if err != nil {
+		log.Print(err)
+		return
+	}
+	session.Values["user_id"] = uid
+	session.Values["csrf_token"] = SecureRandomStr(16)
+	session.Save(r, w)
+
+	http.Redirect(w, r, "/", http.StatusFound)
+}
+
+func GetLogout(w http.ResponseWriter, r *http.Request) {
+	session := GetSession(r)
+	delete(session.Values, "user_id")
+	session.Options = &sessions.Options{MaxAge: -1}
+	session.Save(r, w)
+
+	http.Redirect(w, r, "/", http.StatusFound)
+}
