+++ webapp/golang/handlers/post.go
@@ -1,0 +1,236 @@
+package handlers
+
+import (
+	"fmt"
+	"io"
+	"log"
+	"net/http"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/catatsuy/private-isu/webapp/golang/models"
+)
+
+const (
+	postsPerPage  = 20
+	ISO8601Format = "2006-01-02T15:04:05-07:00"
+	UploadLimit   = 10 * 1024 * 1024 // 10mb
+)
+
+func GetIndex(w http.ResponseWriter, r *http.Request) {
+	me := utils.GetSessionUser(r)
+
+	results := []models.Post{}
+
+	err := db.Select(&results, "SELECT p.id, p.user_id, p.body, p.mime, p.created_at FROM posts p INNER JOIN users u ON p.user_id = u.id WHERE u.del_flg = 0 ORDER BY p.created_at DESC LIMIT ?", postsPerPage)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	posts, err := makePosts(results, utils.GetCSRFToken(r), false)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	template.Must(template.New("layout.html").Funcs(template.FuncMap{
+		"imageURL": ImageURL,
+	}).ParseFiles(
+		getTemplPath("layout.html"),
+		getTemplPath("index.html"),
+		getTemplPath("posts.html"),
+		getTemplPath("post.html"),
+	)).Execute(w, struct {
+		Posts     []models.Post
+		Me        models.User
+		CSRFToken string
+		Flash     string
+	}{posts, me, utils.GetCSRFToken(r), utils.GetFlash(w, r, "notice")})
+}
+
+func GetPosts(w http.ResponseWriter, r *http.Request) {
+	maxCreatedAt := r.URL.Query().Get("max_created_at")
+	if maxCreatedAt == "" {
+		return
+	}
+
+	t, err := time.Parse(ISO8601Format, maxCreatedAt)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	results := []models.Post{}
+	err = db.Select(&results, "SELECT p.id, p.user_id, p.body, p.mime, p.created_at FROM posts p INNER JOIN users u ON p.user_id = u.id WHERE u.del_flg = 0 AND p.created_at <= ? ORDER BY p.created_at DESC LIMIT ?", t.Format(ISO8601Format), postsPerPage)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	posts, err := makePosts(results, utils.GetCSRFToken(r), false)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	if len(posts) == 0 {
+		w.WriteHeader(http.StatusNotFound)
+		return
+	}
+
+	template.Must(template.New("posts.html").Funcs(template.FuncMap{
+		"imageURL": ImageURL,
+	}).ParseFiles(
+		getTemplPath("posts.html"),
+		getTemplPath("post.html"),
+	)).Execute(w, posts)
+}
+
+func GetPostsID(w http.ResponseWriter, r *http.Request) {
+	pidStr := r.PathValue("id")
+	pid, err := strconv.Atoi(pidStr)
+	if err != nil {
+		w.WriteHeader(http.StatusNotFound)
+		return
+	}
+
+	results := []models.Post{}
+	err = db.Select(&results, "SELECT * FROM `posts` WHERE `id` = ?", pid)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	posts, err := makePosts(results, utils.GetCSRFToken(r), true)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	if len(posts) == 0 {
+		w.WriteHeader(http.StatusNotFound)
+		return
+	}
+
+	p := posts[0]
+	me := utils.GetSessionUser(r)
+
+	template.Must(template.New("layout.html").Funcs(template.FuncMap{
+		"imageURL": ImageURL,
+	}).ParseFiles(
+		getTemplPath("layout.html"),
+		getTemplPath("post_id.html"),
+		getTemplPath("post.html"),
+	)).Execute(w, struct {
+		Post models.Post
+		Me   models.User
+	}{p, me})
+}
+
+func PostIndex(w http.ResponseWriter, r *http.Request) {
+	me := utils.GetSessionUser(r)
+	if !utils.IsLogin(me) {
+		http.Redirect(w, r, "/login", http.StatusFound)
+		return
+	}
+
+	if r.FormValue("csrf_token") != utils.GetCSRFToken(r) {
+		w.WriteHeader(http.StatusUnprocessableEntity)
+		return
+	}
+
+	file, header, err := r.FormFile("file")
+	if err != nil {
+		session := utils.GetSession(r)
+		session.Values["notice"] = "画像が必須です"
+		session.Save(r, w)
+
+		http.Redirect(w, r, "/", http.StatusFound)
+		return
+	}
+
+	mime := ""
+	if file != nil {
+		contentType := header.Header["Content-Type"][0]
+		if strings.Contains(contentType, "jpeg") {
+			mime = "image/jpeg"
+		} else if strings.Contains(contentType, "png") {
+			mime = "image/png"
+		} else if strings.Contains(contentType, "gif") {
+			mime = "image/gif"
+		} else {
+			session := utils.GetSession(r)
+			session.Values["notice"] = "投稿できる画像形式はjpgとpngとgifだけです"
+			session.Save(r, w)
+
+			http.Redirect(w, r, "/", http.StatusFound)
+			return
+		}
+	}
+
+	filedata, err := io.ReadAll(file)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	if len(filedata) > UploadLimit {
+		session := utils.GetSession(r)
+		session.Values["notice"] = "ファイルサイズが大きすぎます"
+		session.Save(r, w)
+
+		http.Redirect(w, r, "/", http.StatusFound)
+		return
+	}
+
+	query := "INSERT INTO `posts` (`user_id`, `mime`, `imgdata`, `body`) VALUES (?,?,?,?)"
+	result, err := db.Exec(
+		query,
+		me.ID,
+		mime,
+		filedata,
+		r.FormValue("body"),
+	)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	pid, err := result.LastInsertId()
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	http.Redirect(w, r, "/posts/"+strconv.FormatInt(pid, 10), http.StatusFound)
+}
+
+func PostComment(w http.ResponseWriter, r *http.Request) {
+	me := utils.GetSessionUser(r)
+	if !utils.IsLogin(me) {
+		http.Redirect(w, r, "/login", http.StatusFound)
+		return
+	}
+
+	if r.FormValue("csrf_token") != utils.GetCSRFToken(r) {
+		w.WriteHeader(http.StatusUnprocessableEntity)
+		return
+	}
+
+	postID, err := strconv.Atoi(r.FormValue("post_id"))
+	if err != nil {
+		log.Print("post_idは整数のみです")
+		return
+	}
+
+	query := "INSERT INTO `comments` (`post_id`, `user_id`, `comment`) VALUES (?,?,?)"
+	_, err = db.Exec(query, postID, me.ID, r.FormValue("comment"))
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	http.Redirect(w, r, fmt.Sprintf("/posts/%d", postID), http.StatusFound)
+}
