+++ webapp/golang/handlers/user.go
@@ -1,0 +1,94 @@
+package handlers
+
+import (
+	"log"
+	"net/http"
+	"strings"
+
+	"github.com/catatsuy/private-isu/webapp/golang/models"
+	"github.com/catatsuy/private-isu/webapp/golang/utils"
+)
+
+func GetAccountName(w http.ResponseWriter, r *http.Request) {
+	accountName := r.PathValue("accountName")
+	user := models.User{}
+
+	err := db.Get(&user, "SELECT * FROM `users` WHERE `account_name` = ? AND `del_flg` = 0", accountName)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	if user.ID == 0 {
+		w.WriteHeader(http.StatusNotFound)
+		return
+	}
+
+	results := []models.Post{}
+
+	err = db.Select(&results, "SELECT `id`, `user_id`, `body`, `mime`, `created_at` FROM `posts` WHERE `user_id` = ? ORDER BY `created_at` DESC", user.ID)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	posts, err := makePosts(results, utils.GetCSRFToken(r), false)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	commentCount := 0
+	err = db.Get(&commentCount, "SELECT COUNT(*) AS count FROM `comments` WHERE `user_id` = ?", user.ID)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+
+	postIDs := []int{}
+	err = db.Select(&postIDs, "SELECT `id` FROM `posts` WHERE `user_id` = ?", user.ID)
+	if err != nil {
+		log.Print(err)
+		return
+	}
+	postCount := len(postIDs)
+
+	commentedCount := 0
+	if postCount > 0 {
+		s := []string{}
+		for range postIDs {
+			s = append(s, "?")
+		}
+		placeholder := strings.Join(s, ", ")
+
+		// convert []int -> []interface{}
+		args := make([]interface{}, len(postIDs))
+		for i, v := range postIDs {
+			args[i] = v
+		}
+
+		err = db.Get(&commentedCount, "SELECT COUNT(*) AS count FROM `comments` WHERE `post_id` IN ("+placeholder+")", args...)
+		if err != nil {
+			log.Print(err)
+			return
+		}
+	}
+
+	me := utils.GetSessionUser(r)
+
+	template.Must(template.New("layout.html").Funcs(template.FuncMap{
+		"imageURL": ImageURL,
+	}).ParseFiles(
+		getTemplPath("layout.html"),
+		getTemplPath("user.html"),
+		getTemplPath("posts.html"),
+		getTemplPath("post.html"),
+	)).Execute(w, struct {
+		Posts          []models.Post
+		User           models.User
+		PostCount      int
+		CommentCount   int
+		CommentedCount int
+		Me             models.User
+	}{posts, user, postCount, commentCount, commentedCount, me})
+}
