+++ webapp/golang/models/types.go
@@ -1,0 +1,47 @@
+package models
+
+import (
+	"time"
+)
+
+type User struct {
+	ID          int       `db:"id"`
+	AccountName string    `db:"account_name"`
+	Passhash    string    `db:"passhash"`
+	Authority   int       `db:"authority"`
+	DelFlg      int       `db:"del_flg"`
+	CreatedAt   time.Time `db:"created_at"`
+}
+
+type Post struct {
+	ID           int       `db:"id"`
+	UserID       int       `db:"user_id"`
+	Imgdata      []byte    `db:"imgdata"`
+	Body         string    `db:"body"`
+	Mime         string    `db:"mime"`
+	CreatedAt    time.Time `db:"created_at"`
+	CommentCount int
+	Comments     []Comment
+	User         User
+	CSRFToken    string
+}
+
+type Comment struct {
+	ID        int       `db:"id"`
+	PostID    int       `db:"post_id"`
+	UserID    int       `db:"user_id"`
+	Comment   string    `db:"comment"`
+	CreatedAt time.Time `db:"created_at"`
+	User      User
+}
+
+type PostResult struct {
+	Post Post
+	Err  error
+}
+
+type CommentResult struct {
+	Comments []Comment
+	Count   int
+	Err     error
+}
